Fixed clang warnings

+++ source/misc/utils.cpp
@@ -11,7 +11,8 @@ vector<double> linspace(double xl, double xh, int n)
 {
   vector<double> res(n,0);
   for(size_t i=0;i<size_t(n);++i)
-    res[i] = xl + (xh-xl)*(double)i/(double)(n-1);
+    res[i] = xl + (xh-xl)*static_cast<double>(i)/
+      static_cast<double>(n-1);
   return res;
 }
 
@@ -27,7 +28,7 @@ vector<double> arange(double x_min, double x_max, double dx)
 double min(vector<double> const& v)
 {
   double res = v[0];
-  for(int i=1;i<(int)v.size();++i)
+  for(size_t i=1;i<v.size();++i)
     res = std::min(res,v[size_t(i)]);
   return res;
 }
@@ -35,7 +36,7 @@ double min(vector<double> const& v)
 double max(vector<double> const& v)
 {
   double res = v[0];
-  for(int i=1;i<(int)v.size();++i)
+  for(size_t i=1;i<v.size();++i)
     res = std::max(res,v[size_t(i)]);
   return res;
 }
